# REST APIs Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: https://todo.swagger.io/v2
  - url: http://todo.swagger.io/v2
...
paths:
...
  /api/users:
    get:
      tags:
        - users
      summary: Get information about the users
      description:
        The available information (passwords excluded) about all the users
        is retrieved. This operation is available only to authenticated users.
      operationId: getUsers
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: Users
  /api/users/{userId}:
    get:
      tags:
        - "users/{userId}"
      summary: Get information about a user
      description:
        The available information (password excluded) about the user specified
        by userId is retrieved.
      operationId: getSingleUser
      parameters:
        - name: userId
          in: path
          description: ID of the user to get
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content: {}
        "404":
          description: Not Found
          content: {}
      security:
        - cookieAuth: []
      x-swagger-router-controller: UsersuserId
  /api/users/authenticator:
    post:
      tags:
        - users/authenticator
      summary: Logs a user in or out
      description:
        "The user who wants to log in or out sends the user data to the\
        \ authenticator which performs the operation. For logout, the operation is\
        \ possible only for authenticated users."
      operationId: authenticateUser
      parameters:
        - name: type
          in: query
          description: The operation type ("login" or "logout")
          required: false
          style: form
          explode: true
          schema:
            type: string
      requestBody:
        description:
          The data of the user who wants to perform log in. For login the
          structure must contain email and password.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "204":
          description: No Content
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: token=abcde12345; Path=/; HttpOnly
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
      x-swagger-router-controller: Usersauthenticator
components:
  schemas:
    User:
      required:
        - email
      type: object
      properties:
        $schema:
          type: string
          description:
            path to the JSON Schema to use for the validation of the JSON
            data structure describing the user
        id:
          type: integer
          description: unique identifier for the user
        email:
          type: string
          description: user email
          format: email
        name:
          type: string
          description: username
        password:
          maxLength: 20
          minLength: 6
          type: string
          description: user password
        self:
          type: string
          format: uri
      additionalProperties: false
      description: schema for User data structures
      example:
        password: password
        $schema: $schema
        name: name
        self: http://example.com/aeiou
        id: 0
        email: ""
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie